Update patch set 1
Patch Set 1:

I would probably replace it with an assert that the number of characters is not more than 4 - using an assert as a sort of super-comment so the reader does not have to read the whole function to know that this is the case.

But I don't really work on V8 these days, so you probably want yangguo or jgruber as reviewers that can actually land this.

Patch-set: 1
CC: Gerrit User 1001642 <1001642@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>