Update patch set 3
Patch Set 3:

PTAL ---

Since the reason for the revert was from CodeStubAssembler stuff I'd appreciate a look from ishell or jgruber.

The bug was asserting that the value parameter passed to AsyncGeneratorResolve() was not a JSPromise object (because it should have been Awaited).

This change adds a new helper for these assertions, which checks if the value is an Smi.

A simpler fix would be to just remove the assertion. Whatever people prefer.

Patch-set: 3
Reviewer: Gerrit User 1162379 <1162379@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Work-in-progress: false