Added new explicit constructor taking isolate, instead of using Isolate::Current
As part of J2V8 development (https://github.com/eclipsesource/J2V8), we realized that we had a subtle bug in how Isolate scope was created.
Most likely bug was fixed (currently testing), however, what we did notice as well is that V8 API over time is moving in more and more places
to notion in which one can pass Isolate explicitly to methods or constructors. One such place is missing and we needed this. This contribution
adds explicit Isolate passed to some of constructors in the API.

R=adamk@chromium.org

Bug:
Cq-Include-Trybots: master.tryserver.chromium.linux:linux_chromium_rel_ng
Change-Id: I61984285f152aba5ca922100cf3df913a9cb2cea