[turbofan] Initial support for inlining some Array subclasses.
Refactor the JSCreateArray lowering a bit to allow for some Array
subclasses to be inlined. This doesn't yet deal well with subclasses
that use the default constructor or contain a manual spread super
call. Handling this depends on the Array constructor and the spread
construct builtins, which don't pass an AllocationSite.

Bug: v8:6399

Change-Id: I695e21c821dee70ca44d89ef13f5231c7a2ee7db