Update patch set 10
Patch Set 10:

> Patch Set 9:
> 
> (12 comments)
> 
> > Patch Set 9:
> > 
> > (5 comments)
> > 
> > I like this approach, seems like the way to go to me. Gotta admit I'm not 100% familiar with catch prediction though.
> > 
> > Your last sentence in the previous comment appears to be incomplete :)
> 
> I thought I had erased that entire sentence --- from reading the code, it looked like ResumeGeneratorTrampoline only ever dealt with bytecode, based on
> 
> ```
>     __ movp(rcx, FieldOperand(rdi, JSFunction::kSharedFunctionInfoOffset));
>     __ movp(rcx, FieldOperand(rcx, SharedFunctionInfo::kFunctionDataOffset));
>     __ CmpObjectType(rcx, BYTECODE_ARRAY_TYPE, rcx);
>     __ Assert(equal, kMissingBytecodeArray);
> ```
> 
> but of course, the SharedFunctionInfo always holds the unoptimized code, which should always be interpreted for generators. It's just confusion due to a new awareness that SFI holds unoptimized code apparently always.

No the ResumeGeneratorTrampoline can jump to optimized code. E.g., it jumps the the code pointed to by the JSFunction - see: 
   __ movp(rcx, FieldOperand(rdi, JSFunction::kCodeOffset));
which could point at either the interpreter entry trampoline (unoptimized) or optimized code. Even if it points at the interpreter entry trampoline, that trampoline can "self-heal" the JSFunction to point to optimized code if it sees there is valid optimized code in the feedback vector.

The code you linked above is just ensuring that the unoptimized code is actually bytecode not FCG (but this has always been the case for a long time for Generators, even before I+TF shipped).

Patch-set: 10