Update patch set 1
Patch Set 1:

One of the followups discussed was to avoid inlining the Await operation in so many builtins.

This CL does that (although it also introduces several other changes which I had wanted to make as well).

The snapshot-size savings are not super significant, 5.144kb in an x64.release build, so it may not be entirely worth it. But feel free to take a look.

There is a minor increase in bytecode size for getting rid of the duplicate Caught|Uncaught builtins --- however, with the Await not being inlined in these stubs, the presence of multiple stubs is less important.

Patch-set: 1
Reviewer: Gerrit User 1162379 <1162379@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Work-in-progress: false