Update patch set 2
Patch Set 2:

> Patch Set 1:
> 
> The bytecode array's age is also reset by the InterpreterEntryTrampoline in generated code. Do you take this into account? Won't there be a race between the marker and mutator.
> 
> Also, why does this need to be a CAS? Surely only one task should end up marking (and incrementing the age) of the array during a GC cycle? If not then couldn't we accidentally age the array more than once in a single GC cycle?

There is indeed a race between the mutator resetting the counter and the concurrent marker incrementing it. The same "race" exists with sequential case too. Depending on the order of the marking step on the main thread and the bytecode execution, the counter value at the end of marking will be either 0 or 1.

We need the CAS to preserve the sequential behavior in this mutator/marker "race". Without CAS the following scenario is possible:
1) concurrent marker loads age = kLastBytecodeAge - 1.
2) concurrent marker is descheduled.
3) bytecode runs and sets the age to 0.
4) concurrent marker runs and sets the age to kLastBytecodeAge.

With the CAS the age increment happens completely before or completely after the reset.

> Surely only one task should end up marking (and incrementing the age) of the array during a GC cycle? 
Yes, there is only one task marking. The CAS is for mutator/marker race.

Patch-set: 2
Reviewer: Gerrit User 1002405 <1002405@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>