Update patch set 1
Patch Set 1: Code-Review+1

LGTM. As noted on the linked bug, this runtime function is actually _not_ on the whitelist. But I agree that is would make sense to have this one on the whitelist and be fuzzed, as it might flush out interesting cases with respect to TurboFan inlining. We should add it to the whitelist.

Patch-set: 1
Reviewer: Gerrit User 1118575 <1118575@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=+1