Update patch set 7
Patch Set 7:

> Patch Set 7:
> 
> > Patch Set 7:
> > 
> > Hmm, I think I am missing a lot more checks than whether from_index is a SMI (e.g., adapting a negative from_index).
> > 
> > These checks are all already present in the beginning of ArrayIncludesIndexofAssembler::Generate. Would it make sense to separate them into a yet another stub, and call this stub before one of the existing three stubs?
> 
> Depending on how much code they produce, it's fine to just inline them into the three specialized stubs. 
> 
> I haven't looked at the code in detail, but it sounds like a helper function (used by ArrayIncludesIndexofAssembler::G and by the new builtins) might be the way to go.

Thanks, Jakob!

I'm also adding Benedikt, who suggested this splitting of the builtin, to Cc with a question:

The original motivation for the split of the monolithic builtin was to remove duplicated checks when using the builtin from Turbofan. Some of the checks outside of the three introduced stubs are apparently not duplicated, e.g., the check that from_index is indeed just SMI, or handling the case when it is negative. Checking that from_index is SMI cannot easily be done in the reducer, as Jakob tells me, so I'm trying to figure out which checks I need to add to the stubs.

What are the duplicated checks which do not need to be done in the reducer?

Cheers,
Vaclav

Patch-set: 7
CC: Gerrit User 1183894 <1183894@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Reviewer: Gerrit User 1002541 <1002541@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=0