[modules] Introduce ScriptRecordOrModule
This patch introduces a new sum type for representing both scripts and
modules to be used with dynamic import. ScriptRecord contains the
necessary metadata required by the HTML spec to load a module
dynamically. This is plumbed through the ScriptOrigin.

The dynamic import API is also updated to use this new type.

Bug: v8:5785
Change-Id: I9b4ecd1696a399632b1a21bcea53380d7bdd073c