Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> Hi Adithya,
> 
> Humm, I'm not sure we really need this change. I know currently the situation is not entirely perfect.
> 
> The current solution was based on the limitation that we don't get any timer information from blink, aka, we do have to manually ensure we always pass through a .*Callback.* scope to ensure "proper" Blink C++ time classification.
> Note that we also disable several optimizations to make sure we *always* do another indirection before we actually invoke the callback (which could be done directly in many cases).
> 
> If Blink has all callbacks properly annotated we don't need any parent guard timer scope which only causes overhead in this case. We do need this though in the stand-alone case / where we don't have a custom chrome build with all callback timer scopes in place.

Wouldn't you still need the timer scopes in V8 in order to stop counting time spent in Blink as part of V8's execution time (even if you were to call into Blink directly and Blink had all the callback timer scopes in place?). 

> 
> WDYT?

This change doesn't really improve things too much for RCS measurements on most webpages. It only has a pronounced effect in Speedometer (specifically React) where > 95% of the calls to AccessorNameGetterCallback and GenericNamedSetterCallback actually end up in V8 as opposed to Blink, adding a sizable overhead because a new timer scope is entered when it calls an accessor function. So really it just increases the total % of Blink C++ column in tracing to be closer to 90% for React. So I'm fine with not landing this if you think it adds too much complexity (I can just use this locally to get more accurate percentages).

Patch-set: 4
Reviewer: Gerrit User 1175471 <1175471@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>