Update patch set 10
Patch Set 10:

By the way, for spec context: We discussed this proposal at the July TC39 meeting, but it didn't get consensus, since some committee members were OK with a leak. I plan to bring this up in the September meeting again--I'm not sure if all committee members understood that this is the only leak from evaluating code that is later collected, so I'll try to explain that better this time.

Patch-set: 10
Reviewer: Gerrit User 1129576 <1129576@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>