Update patch set 10
Patch Set 10:

> Patch Set 10: Code-Review-1
> 
> > Patch Set 10:
> > 
> > > Patch Set 8:
> > > 
> > > (1 comment)
> > 
> > Having SFI reference a JSArray is a bad idea. We have several places that assume SFIs to br context-independent. As you have found, the compilation cache (the in memory cache). But also for the custom startup snapshot we have the same issue: SFIs are serialized by the startup serializer, which also cannot deal with anything context-dependent.
> > 
> > Would it be possible to store this JSArray onto the feedback vector?
> 
> Adding the FixedArray in the SFI to store a JSArray can lead to cross context leaks. Two iframes loading the same script containing template literals, the second one would get access to a JSArray from the first iframe.

Okay, so we can't guarantee that SharedFunctionInfo won't leak to different contexts in Chromium. That answers one question.

I guess the workaround if this spec change goes through is to do much the same thing as the current solution --- a WeakMap mapping Context to cached JSArray instances, attached to the SharedFunctionInfo (as opposed to a NativeContext-attached WeakMap which maps template IDs to JSArrays).

Does that sound like a reasonable approach for the proposed spec change, or is there a better option that jumps out?

Patch-set: 10
Reviewer: Gerrit User 1162379 <1162379@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>