[Turbofan] Removing a minor TODO in the OSR part of the BytecodeGraphBuilder
The TODO was about wrapping together the sourceposition iterator and the bytecode
iterator. Since the first one is useful in fewer parts than the second, and the
bytecode iterator is more flexible to advance than the sourceposition iterator,
and we would not gain that much more readability, this TODO is removed.

Bug:
Change-Id: I104d0f5f0cd01686ea48d209419bd6bb2ed19bcf
TBR=mstarzinger@chromium.org