Update patch set 3
Patch Set 3:

(1 comment)

> Patch Set 3:
> 
> (1 comment)
> 
> What benefit does this CL provide?
> 
> The rule of thumb is: prefer handlified code, unless operating on raw pointers provides a specific benefit, like a significant performance improvement due to not having to create handles.
> 
> In this case, AFAICS the function in question is only called from handlified code anyway, and doesn't create any handles itself, so it seems you're simply trading the slightly shorter syntax of "JSReceiver::cast()" vs "Handle<JSReceiver>::cast()" for a bunch of manual handle derefs. Am I overlooking anything?
> 
> If you insist on refactoring this, then why is the function still static, as opposed to being a member function operating on |this|?

In builtins-collections-gen.cc we call out to the runtime function GetExistingHash to create the hash (which in turn calls out to GetOrCreateHash). With this CL, we can change this to be a CallCFunction instead of a runtime call. 

I'm happy to change this to a member function.

Patch-set: 3
Reviewer: Gerrit User 1158954 <1158954@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=0