Update patch set 1
Patch Set 1:

Couple of questions:
 - is must_use_preparsed_scope_data_ only set for non-top level functions (or eagerly compiled inner functions from the top level code)?
 - is there a good way to tell from compiler.cc that there are scopes that must use prepared scope data?

If both are yes then this looks good. It's another thing we will need to deal with if we want to lazy compile off-thread, but we already have that issue with scope infos.

Patch-set: 1
Reviewer: Gerrit User 1002763 <1002763@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>