[wasm] Fix Memory.Grow with guard pages enabled to adjust memory allocated
BUG=5886

Memory.Grow with guard pages enabled should adjust amount of allocated memory,
and not allocate a new buffer. This was disabled because the in the previous
guard pages implementation, the backing store was freed in the MemoryFinalizer,
and we needed to be sure that the backing store is not released till the last
buffer using it is released. This is now safe with --enable-guard region  as
we no longer use the MemoryFinalizer to free the backing store.

Change-Id: I0d7edb884bd1e3167eb5fbced6953c6401688d40