[wasm] Introduce the WasmContext in WasmMemoryObject
The WasmContext struct introduced in this CL is used to store the
mem_size and mem_start address of the wasm memory. These variables can
be accessed at C++ level at graph build time (e.g., initialized during
instance building).  When the WebAssembly grow_memory instruction has
to grows the memory, the GrowMemory runtime is invoked, so we can
directly change the variables in the WasmContext at C++ level.

When the values are needed in the generated code (i.e., for access
bounds checks), at graph creation time we generate
RelocatableIntPtrConstant nodes pointing to the values stored in the
WasmContext in order to load the requested data.  When the memory
changes (instance initialization, reset, grow_memory), then the context
values are changed, and the generated code will load the correct values
from the WasmContext.

This removes the need of patching the generated code at runtime (i.e.,
when the memory grows) with respect to WASM_MEMORY_REFERENCE and
WASM_MEMORY_SIZE_REFERENCE. However, we still need to patch the code at
instance build time (and at compiled module reset time); in fact the
address of the WasmContext is not known during compilation, but only
when the instance is built.

This CL introduces the WasmContext struct, created and attached to every
WasmMemoryObject, this way we can share the same WasmContext with
multiple instances sharing the same wasm memory. In order to always have
a valid WasmContext, we always create a WasmMemoryObject (or reference
an existing one when the memory is imported from another module), for
each module that uses memory. The code needed to patch
WASM_MEMORY_REFERENCE and WASM_MEMORY_SIZE_REFERENCE has been removed in
favor of the one needed to patch WASM_CONTEXT_REFERENCE (the address of
the WasmContext).

R=ahaas@chromium.org,clemensh@chromium.org

Change-Id: I221fc82948d0a375db88ddea5603310b1779926f