[profiling] Add common flag to disable fast/inlined api calls
When profiling, tracing or runtime stats is active we have to disable fast api
calls. Previously we special-cased runtime stats, but the CallApiCallbackStub
already supports a slow-fallback when profiling is active.

This CL adds a separate use_slow_api_callback_address flag on the isolate to
disable these fast paths.

Change-Id: I09bfa50cc87e7294ede391b39bb771af58231093