jumbo: rename struct Register to struct AsmRegister
This avoids conflicts with the Register class defined in the
v8::internal::interpreter namespace, and it can also help to
avoid confusion for human beings over which type a particular
object is.
    
With this change, non-default jumbo_file_merge_limit values
should hopefully work now.
    
This change has only been tested with x64 and arm builds so far.

BUG=chromium:746958

Change-Id: I3f672a9f88501519a25c908823ca71118ca4c950