[DONE] [wasm] Store the WasmContext nodes in the SsaEnv
Storing the WasmContext variables along with the WasmMemoryObject
requires to generate new TF nodes every time we load a context variable
(i.e., every time we need to load/store variables in memory).
We cannot adopt the same strategy that was adopted before
(RelocatableIntPtrConstant nodes cached in WasmGraphBuilder) since the
values represented by the nodes might change only for particular
execution (e.g., grow_memory in a branch), so the cached node could
result in not being in the dominator path when accessed a second time.

In order to avoid loading the context variables when not necessary, in
this CL we store the nodes representing the WasmContext variables
directly in the SSA environment (similar to other local variables that
might change during execution) and use the SSA environment as a cache.

The nodes are created when initializing the SSA environment and
refreshed every time a grow_memory happens, so that we are sure that
they always represent the correct mem_size and mem_start variables. When
validating the wasm code in fuction-body-decoder.cc we also make sure
that the user code is not able to access these special local variables.

This CL also removes the WasmMemorySize runtime (since it's now possible
to directly retrieve mem_size from the context) and simplifies the
GrowMemory runtime (since every instance now has a memory_object).

R=ahaas@chromium.org,clemensh@chromium.org

Change-Id: I4d90b4ace3e6d5baca3710eabc97aa91d57b07f6