Update patch set 1
Patch Set 1:

I'm not sure who is the right person to ask to take a look at this, but it's small.

Seems to fix a TODO which has been in the tree for a very long time, but it's possible that the dummy value was needed for stack alignment or something, I dunno. In either case tests seem to pass without it.

Patch-set: 1
CC: Gerrit User 1118629 <1118629@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
CC: Gerrit User 1183894 <1183894@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
CC: Gerrit User 1118373 <1118373@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Reviewer: Gerrit User 1162379 <1162379@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Work-in-progress: false