[cctest] enable disable_automatic_dispose_ check
Currently there is a TODO in the main function of cctest.cc just before
the commented out line with the disable_automatic_dispose_ check. The
comment refers to another comment above but I've not been able to
understand which comment it is referring to.

Hopefully someone can provide some input regarding the proper way to
handle this. This commit suggests removing the comment and also
enabling the check again.

Bug:
Change-Id: I4367e55f3476e58b08b4f447b4d37955fe1beff9