Update patch set 9
Patch Set 9:

(5 comments)

Looks good, with a few comments.

High-level meta-comment: this CL is bigger than I was expecting. In particular, it contains some amount of throwaway work (anything that's based on the assumption that we have the "BigInt::value()" accessor, or that BigInts have a fixed BigInt::kSize, will have to be changed again). My suggestion would have been to limit the first CL to basic infrastructure that's unlikely to change, even if that means it's not testable right away. Oh well, now we'll build on this :-)

Patch-set: 9
Reviewer: Gerrit User 1115935 <1115935@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>